functions/requirements.txt
Flask>=2.0.1,<3.0.0
Pillow>=9.5.0,<10.0.0
requests>=2.28.0,<3.0.0
python-dotenv>=0.19.0,<1.0.0
Werkzeug>=2.0.1,<3.0.0 